scraping-service/requirements.txt
geocoder
requests
selenium
python-dotenv
undetected-chromedriver
ijson
aiohttp
asyncio
tqdm
